# src/MX.TripSideKick.Web/ClientApp/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "MX.TripSideKick.Web | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost/"
    }
  ],
  "paths": {
...
    "/v1/auth/login": {
      "get": {
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "name": "returnUrl",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": { }
            }
          }
        },
        "security": [
          {
            "SessionCookie": [ ]
          }
        ]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "name": "returnUrl",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": { }
            }
          }
        },
        "security": [
          {
            "SessionCookie": [ ]
          }
        ]
      }
    },
    "/v1/auth/antiforgery": {
      "get": {
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": { }
            }
          }
        },
        "security": [
          {
            "SessionCookie": [ ]
          }
        ]
      }
    },
    "/v1/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": [ ]
          }
        ]
      }
    },
...
    "/v1/trips/{tripId}/invitations": {
      "get": {
        "tags": [
          "Invitations"
        ],
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvitationResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": [ ]
          }
        ]
      },
      "post": {
        "tags": [
          "Invitations"
        ],
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvitationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvitationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvitationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": [ ]
          }
        ]
      }
    },
    "/v1/trips/{tripId}/invitations/{invitationId}/resend": {
      "post": {
        "tags": [
          "Invitations"
        ],
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "invitationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": [ ]
          }
        ]
      }
    },
    "/v1/trips/{tripId}/invitations/{invitationId}/revoke": {
      "post": {
        "tags": [
          "Invitations"
        ],
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "invitationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": { }
            }
          }
        },
        "security": [
          {
            "SessionCookie": [ ]
          }
        ]
      }
    },
    "/v1/invitations/accept": {
      "post": {
        "tags": [
          "Invitations"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvitationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvitationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvitationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcceptInvitationRequest": {
        "required": [
          "acceptanceToken"
        ],
        "type": "object",
        "properties": {
          "acceptanceToken": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "AuthMeResponse": {
        "required": [
          "isAuthenticated",
          "displayName",
          "subjectId"
        ],
        "type": "object",
        "properties": {
          "isAuthenticated": {
            "type": "boolean"
          },
          "displayName": {
            "type": [
              "null",
              "string"
            ]
          },
          "subjectId": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "CreateInvitationRequest": {
        "required": [
          "invitedEmail",
          "role",
          "linkKind",
          "existingTravellerId",
          "newTravellerDisplayName"
        ],
        "type": "object",
        "properties": {
          "invitedEmail": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/MembershipRole"
          },
          "linkKind": {
            "$ref": "#/components/schemas/TravellerLinkKind"
          },
          "existingTravellerId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "newTravellerDisplayName": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "InvitationResponse": {
        "required": [
          "id",
          "tripId",
          "invitedEmail",
          "role",
          "status",
          "linkKind",
          "acceptanceUrl"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tripId": {
            "type": "string",
            "format": "uuid"
          },
          "invitedEmail": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/MembershipRole"
          },
          "status": {
            "type": "string"
          },
          "linkKind": {
            "$ref": "#/components/schemas/TravellerLinkKind"
          },
          "acceptanceUrl": {
            "type": "string"
          }
        }
      },
...
      "MembershipResponse": {
        "required": [
          "id",
          "tripId",
          "subjectId",
          "role",
          "eTag"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tripId": {
            "type": "string",
            "format": "uuid"
          },
          "subjectId": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/MembershipRole"
          },
          "eTag": {
            "type": "string"
          }
        }
      },
      "MembershipRole": {
        "type": "integer"
      },
...
      "TravellerLinkKind": {
        "type": "integer"
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "description": "Same-origin session cookie established by the OpenID Connect sign-in flow. Mutating requests additionally require the X-CSRF-TOKEN header (see /v1/auth/antiforgery).",
        "name": ".AspNetCore.Cookies",
        "in": "cookie"
      }
    }
  },
  "security": [
    {
      "SessionCookie": [ ]
    }
  ],
...
}